api/requirements.txt
pymongo==3.12.0
Flask_PyMongo==2.3.0
Werkzeug==1.0.1
Flask==1.1.2
python-dotenv==0.19.0
